Notification Model
The user who’s made the new comment or like.
The user who owns the post that has been affected.
The post which the user is commenting or liking.
The id of the comment.
The type of notification it is, a comment or like.
Whether or not the notification has been read.